refactor(autosave): move autosave machinery out of the base package into Nucs.JsonSettings.Autosave
The base Nucs.JsonSettings.dll shipped the whole Nucs.JsonSettings.Autosave namespace as public API, but for a regular settings class autosave is entirely the Autosave package's woven path -- so [IgnoreAutosave], NotificationBinder and the reflection opt-in rules did nothing with the base package alone. This moves the parts base does not itself use into the package where they are actually wired up.

Moved to Nucs.JsonSettings.Autosave (namespace unchanged, so consumers of the package are unaffected):
- IgnoreAutosaveAttribute, NotificationBinder, and the SuspendAutosave struct (git-moved).
- The property opt-in predicates (IsAutosaveOptedIn / IsVersionableVersion / IsAutosaveMonitored / IsNotificationBindable) and _frameworkParameters, into a new internal AutosavePolicy.

Kept in base as the slim shared save-suspension core: AutosaveModule and the AutosavingState enum. SettingsBag's dictionary autosave and JsonSettings.LoadJson drive this state with the base package alone, and the woven path reuses the same instance, so it cannot leave base. Consequences of the move:
- AutosaveModule.SuspendAutosave() is removed; the settings.SuspendAutosave() extension now constructs the struct directly (new SuspendAutosave(module)).
- AutosaveModule.NotificationsHandler is retyped NotificationBinder? -> IDisposable? so base no longer references the moved binder; the module still owns the binder's lifetime and disposes it.

Ripple: AutosaveRuntime and NotificationBinder now call AutosavePolicy; NotifyChangesMixinAttribute's doc cref to NotificationBinder was de-linked because the NotifyChanges package does not reference Autosave; one test gains a (NotificationBinder) cast and one test cref was updated. Recorded in the 2.2.0 PackageReleaseNotes.

Verified: full solution compiles on netstandard2.0/net48/net6.0/net8.0/net10.0; test suite green on every framework (458 passed on net6/8/10, 453 on net472/net48 -- the delta is .NET6+-only AEAD tests -- 0 failed).
