+++ src/JsonSettings.Autosave/IgnoreAutosaveAttribute.cs
