+++ src/JsonSettings.Autosave/SuspendAutosave.cs
